requirements.txt
python=3.9.12
numpy=1.22.3
scikit-learn=1.0.2
yacs=0.1.6
pytorch=1.11.0
torchvision=1.12.0
torchaudio=1.11.0
cudatoolkit=11.3
torch-scatter=0.6.13
torch-sparse=2.0.9
torch-geometric-temporal=0.53.0
pyg=2.0.4
deeprobust=0.2.4
tqdm